chore: fixes camera delete
