+++ src/util/get-camera-service-for-legacy-camera-type.ts
@@ -1,0 +1,14 @@
+import { CameraService, LegacyCameraType } from '@/store/cameras/types'
+
+const getCameraServiceForLegacyCameraType = (type: LegacyCameraType): CameraService => {
+  switch (type) {
+    case 'mjpgstream':
+      return 'mjpegstreamer'
+    case 'mjpgadaptive':
+      return 'mjpegstreamer-adaptive'
+    default:
+      return type
+  }
+}
+
+export default getCameraServiceForLegacyCameraType
